feat: Applying super admin permission
- to allocate admin
- to update connections
- to create / edit recommendation

TISNEW-5817
